+++ src/Services/races-service.js
@@ -1,0 +1,7 @@
+const RacesService = {
+  getAllRaces(knex) {
+    return knex.select('*').from('races_data');
+  },
+};
+
+module.exports = RacesService;
